Completed CDT

+++ RocaloidEngine/LibCyberBase/RDLIO.h
@@ -24,7 +24,7 @@
 {
 	#define Trim(x) trim(x)
 	int TestIfIsInt(string _String);
-	int TestIfIsDouble(string _String);
+	double TestIfIsDouble(string _String);
 	double TestIfIsDoubleAndPositive(string _String);
 	int TestIfIsIntAndPositive(string _String) ;
 	double TestIfIsDoubleNotNegative(string _String);
